# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: OTP Service API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /otp/generate:
    post:
      tags: [OTP]
      summary: Generate OTP code for operation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateOtpRequest'
      responses:
        '200':
          description: Generated OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpResponse'
  /otp/validate:
    post:
      tags: [OTP]
      summary: Validate OTP code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateOtpRequest'
      responses:
        '200':
          description: OTP validation result
          content:
            application/json:
              schema:
                type: string
                example: OTP valid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GenerateOtpRequest:
      type: object
      required: [operationId]
      properties:
        operationId:
          type: integer
          example: 1001
    ValidateOtpRequest:
      type: object
      required: [operationId, code]
      properties:
        operationId:
          type: integer
        code:
          type: string
          example: "123456"
    OtpResponse:
      type: object
      properties:
        code:
          type: string
          example: "123456"
...